requirements.txt
streamlit==1.37.1
faiss-cpu==1.8.0
pypdf==4.2.0
python-docx==1.1.2
pandas==2.2.2
requests==2.32.3
tqdm==4.66.4
